+++ scripts/weaponize.sh
@@ -9,7 +9,7 @@ if [ -e $1 ]; then
 	done;
 
 	if [ $I -gt 2 ]; then
-		echo "At least one file exists on the directory. Exiting.";
+		echo "At least one file exists in the directory. Exiting.";
 	else
 
 		if [ "" == "$3" ]; then
@@ -22,7 +22,7 @@ if [ -e $1 ]; then
 
 			if [ $? != 0 ]; then
 
-				echo "some error has occurr!!!Exiting!!";
+				echo "some error has occurred!!! Exiting!!!";
 				exit;
 
 			fi;
@@ -45,7 +45,7 @@ if [ -e $1 ]; then
 
 			else
 
-				echo "An error has occurr!!!"
+				echo "An error has occurred!!!"
 
 			fi;
 
@@ -54,7 +54,6 @@ if [ -e $1 ]; then
 	fi;
 
 else
-	echo "The odt file does not exists!!";
+	echo "The odt file does not exist!!";
 
 fi;
-
